+++ src/ast/passes/codegen_llvm.cpp
@@ -303,7 +303,7 @@ void CodegenLLVM::kstack_ustack(const std::string &ident,
         b_.CreateGEP(stack_struct, buf, { b_.getInt64(0), b_.getInt32(1) }));
     // store pid
     b_.CreateStore(
-        b_.CreateGetPid(loc),
+        b_.CreateGetPid(ctx_, loc),
         b_.CreateGEP(stack_struct, buf, { b_.getInt64(0), b_.getInt32(2) }));
     // store probe id
     b_.CreateStore(
@@ -346,9 +346,9 @@ void CodegenLLVM::visit(Builtin &builtin)
   } else if (builtin.ident == "kstack" || builtin.ident == "ustack") {
     kstack_ustack(builtin.ident, builtin.type.stack_type, builtin.loc);
   } else if (builtin.ident == "pid") {
-    expr_ = b_.CreateGetPid(builtin.loc);
+    expr_ = b_.CreateGetPid(ctx_, builtin.loc);
   } else if (builtin.ident == "tid") {
-    expr_ = b_.CreateGetTid(builtin.loc);
+    expr_ = b_.CreateGetTid(ctx_, builtin.loc);
   } else if (builtin.ident == "cgroup") {
     expr_ = b_.CreateGetCurrentCgroupId(builtin.loc);
   } else if (builtin.ident == "uid" || builtin.ident == "gid" ||
@@ -400,7 +400,7 @@ void CodegenLLVM::visit(Builtin &builtin)
     }
 
     if (builtin.type.IsUsymTy()) {
-      expr_ = b_.CreateUSym(expr_, get_probe_id(), builtin.loc);
+      expr_ = b_.CreateUSym(ctx_, expr_, get_probe_id(), builtin.loc);
       Value *expr = expr_;
       expr_deleter_ = [this, expr]() { b_.CreateLifetimeEnd(expr); };
     }
@@ -431,7 +431,7 @@ void CodegenLLVM::visit(Builtin &builtin)
       expr_ = b_.CreateRegisterRead(ctx_, builtin.ident);
 
     if (builtin.type.IsUsymTy()) {
-      expr_ = b_.CreateUSym(expr_, get_probe_id(), builtin.loc);
+      expr_ = b_.CreateUSym(ctx_, expr_, get_probe_id(), builtin.loc);
       Value *expr = expr_;
       expr_deleter_ = [this, expr]() { b_.CreateLifetimeEnd(expr); };
     }
@@ -1022,7 +1022,7 @@ void CodegenLLVM::visit(Call &call)
     auto scoped_del = accept(call.vargs.front());
   } else if (call.func == "usym") {
     auto scoped_del = accept(call.vargs.front());
-    expr_ = b_.CreateUSym(expr_, get_probe_id(), call.loc);
+    expr_ = b_.CreateUSym(ctx_, expr_, get_probe_id(), call.loc);
   } else if (call.func == "ntop") {
     // struct {
     //   int af_type;