+++ src/bpftrace.cpp
@@ -2069,6 +2069,20 @@ std::unordered_set<std::string> BPFtrace::get_func_modules(
 #endif
 }
 
+const struct stat &BPFtrace::get_pidns_self_stat() const
+{
+  static struct stat pidns = []() -> auto {
+    struct stat s;
+    if (::stat("/proc/self/ns/pid", &s))
+      throw std::runtime_error(
+          std::string("Failed to stat /proc/self/ns/pid: ") +
+          std::strerror(errno));
+    return s;
+  }();
+
+  return pidns;
+}
+
 Dwarf *BPFtrace::get_dwarf(const std::string &filename)
 {
   auto dwarf = dwarves_.find(filename);