+++ src/bpftrace.h
@@ -9,6 +9,7 @@
 #include <optional>
 #include <set>
 #include <string_view>
+#include <sys/stat.h>
 #include <unordered_map>
 #include <utility>
 #include <vector>
@@ -157,6 +158,8 @@ public:
   virtual bool is_traceable_func(const std::string &func_name) const;
   virtual std::unordered_set<std::string> get_func_modules(
       const std::string &func_name) const;
+  virtual const struct stat &get_pidns_self_stat() const;
+
   bool write_pcaps(uint64_t id, uint64_t ns, uint8_t *pkt, unsigned int size);
 
   void parse_btf(const std::set<std::string> &modules);